feat(ble): complete protocol audit, sensor redesign, and reliability fixes
Major refactor of BLE layer and Sensor UI based on btsnoop log analysis and Gadgetbridge reference.

Protocol & Connection Reliability:
- Aligned connection/init sequence with Gadgetbridge (introduced critical delays, removed invalid hacks).
- Implemented robust "Auto-Redirect to Dashboard" on disconnection events.
- Corrected manual command packet formatting (fixed raw bytes vs. packet structure issue).
- Analyzed btsnoop logs to confirm correct command IDs for Real-Time data (`0x69 01` HR, `0x69 08` Stress/SpO2).

Sensor Functionality:
- Added HRV (Heart Rate Variability) metric support:
  - Implemented `0x69 0A` parsing for Real-Time HRV.
  - Added HRV metric to Dashboard and Sensor Screen.
- Fixed Stress Test Logic:
  - Switched from background monitor command (`0x36 01`) to correct Real-Time request (`0x69 08`).
  - Corrected parsing logic to identify `0x69 08` as Stress (was previously misidentified as SpO2).

UI & UX Improvements:
- Redesigned Sensor Screen Manual Controls:
  - Replaced buttons with a modern 2x2 Grid of interactive Tiles.
  - Added visual states: Colors (Red/Blue/Teal/Purple) and "ACTIVE" indicators.
  - Improved touch feedback with `InkWell/Material`.
- Added "Auto-Reset" logic:
  - Sensor buttons (HR, SpO2, HRV, Stress) now auto-reset to "inactive" if no data is received for 3 seconds (silence detection).
- Added user feedback (SnackBar) when attempting interactions while disconnected.

History & Parsing:
- Fixed timestamp parsing for historical logs to prevent "Future Data" errors.
- Implemented explicit date parsing for Steps (`0x43`) and time-correction for HR/SpO2 logs.

+++ lib/history_screen.dart
@@ -1,3 +1,4 @@
+import 'dart:math'; // For Point
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -153,7 +154,7 @@ class _StepsChartPageState extends State<_StepsChartPage> {
     int currentTotal = 0;
     for (int i = 0; i < sortedSteps.length; i++) {
       final point = sortedSteps[i];
-      currentTotal += point.y;
+      currentTotal += point.y.toInt();
       spots.add(FlSpot(point.x.toDouble(), currentTotal.toDouble()));
     }
